+++ workspaces/ui-v2/src/optic-components/pages/docs/components/index.ts
@@ -1,0 +1,2 @@
+export * from './DocPageNavigation';
+export * from './EditContributionsButton';
