+++ workspaces/ui-v2/src/optic-components/pages/docs/index.ts
@@ -1,0 +1,1 @@
+export * from './DocumentationPages';
