+++ workspaces/ui-v2/src/spectacle-implementations/cloud-viewer.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { useRouteMatch, useParams, Switch } from 'react-router-dom';
 import { Provider as BaseUrlProvider } from '../optic-components/hooks/useBaseUrl';
 import { makeSpectacle, SpectacleInput } from '@useoptic/spectacle';
-import { DocumentationPages } from '../optic-components/pages/docs/DocumentationPage';
+import { DocumentationPages } from '<src>/optic-components/pages/docs';
 import { SpectacleStore } from './spectacle-provider';
 import { Loading } from '../optic-components/loaders/Loading';
 import { DiffReviewEnvironments } from '../optic-components/pages/diffs/ReviewDiffPages';
